Added admixture plot indfile and indfile maker
Although the code is not the best and is not nearly as perfected as the other files I am uploading this code and output. I would also like to thank my friend and colleague Alexandre because without him it would have been impossible for me to finish this is time.